app/Http/Requests/Manager/Studentディレクトリが肥大化しているための整理(yuki)
