+++ app/Http/Requests/Manager/Student/ShowRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Student;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StudentStoreRequest extends FormRequest
+class ShowRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,8 +24,14 @@ class StudentStoreRequest extends FormRequest
     public function rules()
     {
         return [
-            'given_name_by_instructor' => ['required', 'string', 'max:50'],
-            'email' => ['required', 'email', 'max:255', 'unique:students'],
+            'student_id' => ['required', 'integer', 'exists:students,id,deleted_at,NULL'],
         ];
+    }
+
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            'student_id' => $this->route('student_id'),
+        ]);
     }
 }
